+++ src/System/Router/RouteGroup.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace System\Router;
 
 use Closure;
@@ -19,7 +21,9 @@ class RouteGroup
   public function group(Closure $callback)
   {
     call_user_func($this->setup);
-    call_user_func($callback);
+    $result = call_user_func($callback);
     call_user_func($this->cleanup);
+
+    return $result;
   }
 }