+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Change `Router::middleware` param form middleware object to middlewar class-name.
+- `RouterGroup::group` retrun as callback param.
 
 ## [0.6.0] - 2022-05-15
 ### Added